+++ CHANGELOG.md
@@ -16,7 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Methods: 
   - `isAlpha` | `isNotAlpha`
-  - `containsSpecialCharacters` | `notContainsSpecialCharacters`
+  - `containsSpecialCharacter` | `notContainsSpecialCharacter`
 
 ## [1.0.1] - 2023-08-13